Make sanitization aware of alphabets outside basic latin.
Use the XRegExp library to handle letters from the range of unicode alphabets.
Using '\\pL' as a regular expression is equivalent to specifying the unicode
ranges of all alphabets.
The JavaScript toLowerCase function is already unicode aware.
I don't believe that we have to worry about surrogate pairs and their effect
on length, as the alphabets are outside the astral character range.